+++ Assets/HUF/InitFirebase/Editor/FirebaseVersionConsolidator.cs
@@ -1,84 +1,96 @@
 #if HUFEXT_PACKAGE_MANAGER
-using System.Collections.Generic;
-using System.Linq;
+using System.IO;
+using System.Text.RegularExpressions;
+using HUF.Utils.Runtime.Extensions;
+using HUFEXT.PackageManager.Editor.Commands.Data;
 using UnityEditor;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
-using UnityEditor.PackageManager;
-using UnityEditor.PackageManager.Requests;
+using HUFEXT.PackageManager.Editor.Core;
+using HUFEXT.PackageManager.Editor.Models;
 
 public class FirebaseVersionConsolidator : IPreprocessBuildWithReport
 {
-    static readonly string PACKAGE_PREFIX = "com.google.firebase";
-    static readonly string CORE_PACKAGE = "com.google.firebase.app";
+    const string FIREBASE_SDK_UNITY_PLUGIN = "Firebase SDK Unity plugin";
+    const string COM_HUUUGE_PLUGINS_FIREBASE = "com.huuuge.plugins.firebase";
+    const string DOT_FIREBASE = ".firebase";
+    const string ZERO_VERSION = "0.0.0";
+
+    static readonly Regex findVersionInChangelogRegex = new Regex( @"\d*[.]\d*[.]\d*" );
 
     public int callbackOrder => int.MinValue;
 
-    public virtual void OnPreprocessBuild(BuildReport report)
+    public virtual void OnPreprocessBuild( BuildReport report )
     {
-        Execute(true);
+        Execute();
     }
 
-    [MenuItem("HUF/Consolidate Firebase SDKs versions")]
+    [MenuItem( "HUF/Check Firebase SDKs versions" )]
     static void MenuItem()
     {
-        Execute(false);
+        Execute();
     }
 
-    static void Execute(bool check)
+    static void Execute()
     {
-#if UNITY_2019_1_OR_NEWER
-        var listRequest = Client.List(false, true);
-#else
-        var listRequest = Client.List(false);
-#endif
-        EditorApplication.update += ListResponse;
-
-        void ListResponse()
+        Command.Execute( new GetLocalPackagesCommand()
         {
-            if (!listRequest.IsCompleted || listRequest.Status != StatusCode.Success)
-                return;
-
-            EditorApplication.update -= ListResponse;
-            var corePackage = listRequest.Result.ToDictionary(info => info.name)[CORE_PACKAGE];
-            if (corePackage == null)
-                return;
-
-            var packagesToUpdateText = string.Empty;
-            var packagesToUpdate = new Queue<string>();
-            foreach (var package in listRequest.Result)
-                if (package.name.Contains(PACKAGE_PREFIX) && package.version != corePackage.version)
-                {
-                    packagesToUpdate.Enqueue($"{package.name}@{corePackage.version}");
-                    packagesToUpdateText += $"{package.packageId}\n";
-                }
-
-            if (packagesToUpdate.Count == 0)
-                return;
-
-            if (check)
+            OnComplete = ( result, serializedData ) =>
             {
-                EditorUtility.DisplayDialog("IMPORTANT",
-                    $"Firebase SDK packages not matching version {corePackage.version} were found:\n\n{packagesToUpdateText}\n" +
-                    "Please use \"HUF\" -> \"Consolidate Firebase SDKs versions\" menu in order to unify their versions.",
-                    "OK");
-                return;
-            }
+                var localPackages = HUFEXT.PackageManager.Editor.Core.Packages.Local;
 
-            AddRequest addRequest = Client.Add(packagesToUpdate.Dequeue());
-            EditorApplication.update += AddResponse;
+                var corePackage =
+                    localPackages.Find( p => p.name == COM_HUUUGE_PLUGINS_FIREBASE );
 
-            void AddResponse()
-            {
-                if (!addRequest.IsCompleted || addRequest.Status != StatusCode.Success)
+                if ( corePackage == null )
                     return;
 
-                if (packagesToUpdate.Count > 0)
-                    addRequest = Client.Add(packagesToUpdate.Dequeue());
-                else
-                    EditorApplication.update -= AddResponse;
+                var corePackageSDKVersion = GetFirebaseSDKVersionFromChangelog( corePackage );
+                var packagesToUpdateText = string.Empty;
+
+                foreach ( var package in localPackages )
+                {
+                    if ( package.name.EndsWith( DOT_FIREBASE ) )
+                    {
+                        var packageSDKVersion = GetFirebaseSDKVersionFromChangelog( package );
+
+                        if ( packageSDKVersion != corePackageSDKVersion )
+                            packagesToUpdateText += $"{package.displayName}\n";
+                    }
+                }
+
+                if ( packagesToUpdateText.IsNullOrEmpty() )
+                    return;
+
+                EditorUtility.DisplayDialog( "IMPORTANT",
+                    $"Firebase SDK packages not matching version {corePackage.version} were found:\n\n{packagesToUpdateText}\n" +
+                    "Please open HUF Package Manager and install Firebase packages with the same SDK version",
+                    "OK" );
+                return;
+            }
+        } );
+    }
+
+    static string GetFirebaseSDKVersionFromChangelog( PackageManifest package )
+    {
+        var changelogPath = $"{package.huf.path}/CHANGELOG.md";
+
+        foreach ( var line in File.ReadLines( changelogPath ) )
+        {
+            if ( line.Contains( FIREBASE_SDK_UNITY_PLUGIN ) )
+            {
+                var match = findVersionInChangelogRegex.Match( line );
+
+                if ( match.Success )
+                {
+                    return match.Value;
+                }
+
+                break;
             }
         }
+
+        return ZERO_VERSION;
     }
 }
 #endif